+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - added
   - taxel-py: Generate xml from xsd schema
   - Add CI
+  - Add justfile
+  - taxel-gui: Import and display xml
 - changed
   - Update Rust to 1.93
 - removed